Only transform logging result if it's a timeseries
This will probably require some extra work. A previous commit has
already made the min/max fields of the LoggingInfo dataclass optional.
The correct solution for this problem is probably to add a new field to
LoggingInfo which informs defines the type of the logged data. The type
then decides what type of pre/post-processing the data is subjected to.
In particular, the type can make certain optional fields, like min/max
mandatory again.