Change behavior of candriver, fix functional test
Encapsulation of listener from CanBus is weakened in order to allow
users to change the listener more comfortably. This allows creating a
testing setup where the middle of three vcan's passes thru data from one
end to another. This setup is used to test the function of the can
driver.